Add rabbitmq-c.
This is intended to support building the longears package [0].

This commit vendors the PKGBUILD and patch for rabbitmq-c from upstream
commit 22081911686a4c9b37bb3a4364f028f945fdc377 [1]. It modifies the
PKGBUILD to remove the dependency on popt (which is only used for the
optional tools) and build only the static library (disabling the shared
library, tools, examples, and tests, which are irrelevant to R packages
that might use it).

In addition, it introduces a second patch [2] from the unreleased master
branch of the original project, which is needed to enabled building
against pthreads. Without this patch I discovered that the library will
compile correctly but it cannot actually be linked into an R package.

[0]: https://github.com/atheriel/longears
[1]: https://github.com/msys2/MINGW-packages/commit/22081911686a4c9b37bb3a4364f028f945fdc377
[2]: https://github.com/alanxz/rabbitmq-c/commit/6f7f7932f46018925a95118b1173a9aa8d81a50d